# specs/001-product-manager-agent/contracts/api.yaml
openapi: 3.0.3
info:
  title: 产品经理 Agent API
  version: 1.0.0
  description: 智能产品需求分析与PRD生成助手的API契约
servers:
  - url: /api/v1
paths:
...
  /conversations:
    get:
      summary: 获取会话列表
      tags:
        - 会话管理
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: page_size
          in: query
          schema:
            type: integer
            default: 20
      responses:
        '200':
          description: 会话列表
          content:
            application/json:
              schema:
                type: object
                properties:
                  items:
                    type: array
                    items:
                      $ref: '#/components/schemas/Conversation'
                  total:
                    type: integer
                  page:
                    type: integer
                  page_size:
                    type: integer
    
    post:
      summary: 创建新会话
      tags:
        - 会话管理
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                initial_message:
                  type: string
                  description: 用户的第一条消息
      responses:
        '201':
          description: 会话创建成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Conversation'
        '429':
          description: 达到会话创建限制
...
  /admin/llm-configs:
    get:
      summary: 获取所有大模型配置
      tags:
        - 大模型配置（管理员）
      responses:
        '200':
          description: 配置列表
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/LLMConfig'
    
    post:
      summary: 创建大模型配置
      tags:
        - 大模型配置（管理员）
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LLMConfigCreate'
      responses:
        '201':
          description: 配置创建成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LLMConfig'
  /admin/llm-configs/{config_id}:
    get:
      summary: 获取配置详情
      tags:
        - 大模型配置（管理员）
      parameters:
        - name: config_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: 配置详情
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LLMConfig'
    
    put:
      summary: 更新配置
      tags:
        - 大模型配置（管理员）
      parameters:
        - name: config_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LLMConfigCreate'
      responses:
        '200':
          description: 更新成功
    
    delete:
      summary: 删除配置
      tags:
        - 大模型配置（管理员）
      parameters:
        - name: config_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '204':
          description: 删除成功
  /admin/llm-configs/{config_id}/test:
    post:
      summary: 测试配置连接
      tags:
        - 大模型配置（管理员）
      parameters:
        - name: config_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: 连接测试结果
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  message:
                    type: string
                  latency_ms:
                    type: integer
  /admin/llm-configs/{config_id}/set-default:
    post:
      summary: 设置为默认配置
      tags:
        - 大模型配置（管理员）
      parameters:
        - name: config_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: 设置成功
  /admin/llm-stats:
    get:
      summary: 获取大模型调用统计
      tags:
        - 大模型配置（管理员）
      parameters:
        - name: start_date
          in: query
          schema:
            type: string
            format: date
        - name: end_date
          in: query
          schema:
            type: string
            format: date
      responses:
        '200':
          description: 统计信息
          content:
            application/json:
              schema:
                type: object
                properties:
                  total_calls:
                    type: integer
                  avg_latency_ms:
                    type: integer
                  total_tokens:
                    type: integer
                  error_rate:
                    type: number
                    format: float
                  by_provider:
                    type: array
                    items:
                      type: object
                      properties:
                        provider:
                          type: string
                        calls:
                          type: integer

  # ========== 限流配置（管理员） ==========
  /admin/rate-limits:
    get:
      summary: 获取限流配置
      tags:
        - 限流配置（管理员）
      responses:
        '200':
          description: 限流配置
          content:
            application/json:
              schema:
                type: object
                properties:
                  normal_user:
                    type: object
                    properties:
                      hourly_limit:
                        type: integer
                  premium_user:
                    type: object
                    properties:
                      hourly_limit:
                        type: integer
    
    put:
      summary: 更新限流配置
      tags:
        - 限流配置（管理员）
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                normal_user:
                  type: object
                  properties:
                    hourly_limit:
                      type: integer
                premium_user:
                  type: object
                  properties:
                    hourly_limit:
                      type: integer
      responses:
        '200':
          description: 更新成功
components:
  schemas:
    Conversation:
      type: object
      properties:
        id:
          type: string
          format: uuid
        user_id:
          type: string
          format: uuid
        title:
          type: string
        current_stage:
          type: string
          enum: [requirement_clarification, market_research, architecture_design, detailed_design, integration_output]
        version:
          type: integer
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
...
    LLMConfig:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        provider_type:
          type: string
          enum: [openai, anthropic, azure_openai, baidu, aliyun, other]
        base_url:
          type: string
        model_name:
          type: string
        request_params:
          type: object
        is_default:
          type: boolean
        is_active:
          type: boolean
        created_at:
          type: string
          format: date-time
    LLMConfigCreate:
      type: object
      required:
        - name
        - provider_type
        - base_url
        - api_key
        - model_name
      properties:
        name:
          type: string
        provider_type:
          type: string
          enum: [openai, anthropic, azure_openai, baidu, aliyun, other]
        base_url:
          type: string
        api_key:
          type: string
          description: API Key（明文，后端加密存储）
        model_name:
          type: string
        request_params:
          type: object
          properties:
            temperature:
              type: number
            max_tokens:
              type: integer
            timeout:
              type: integer
        is_default:
          type: boolean
        is_active:
          type: boolean